split cylinder into cylinder2d and cylinder3d
