+++ flow360/examples/cylinder2D.py
