+++ flow360/examples/cylinder3D.py
@@ -1,0 +1,13 @@
+"""
+cylinder example
+"""
+
+from .base_test_case import BaseTestCase
+
+
+class Cylinder3D(BaseTestCase):
+    name = "cylinder3D"
+
+    class url:
+        geometry = "local://cylinder.x_t"
+        surface_json = "local://surface_params.json"
