+++ flow360/examples/cylinder3D/surface_params.json
